ai-agent/requirements.txt
livekit==1.0.12
openai==1.3.7
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0